+++ CHANGES.md
@@ -5,6 +5,7 @@
 - feat: for string schemas with kind uuid, check for version to determine regex for validation
 - feat: support schema `type: 'file'` with file type
 - feat: `min` and `max` schema validation for Maps
+- fix: delegate to Zod 4's regexes for string validation
 - fix: remove dropped `unknownKeys`
 - fix: allow symbol again, but given dropped schema support
 - fix: drop nativeEnum as dropped from Zod/Zodexy